FIX: Split functions into different scripts (#5)
* FIX: Split functions into different scripts

* put all scrips in a card dir

* removed two files
