+++ q2_amr/tests/card/test_utils.py
@@ -3,8 +3,8 @@ from unittest.mock import patch
 
 from qiime2.plugin.testing import TestPluginBase
 
+from q2_amr.card.utils import load_preprocess_card_db
 from q2_amr.types import CARDDatabaseFormat
-from q2_amr.utils import load_preprocess_card_db
 
 
 class TestAnnotateReadsCARD(TestPluginBase):
@@ -12,7 +12,7 @@ class TestAnnotateReadsCARD(TestPluginBase):
 
     def test_load_card_db(self):
         card_db = CARDDatabaseFormat()
-        with patch("q2_amr.utils.run_command") as mock_run_command:
+        with patch("q2_amr.card.utils.run_command") as mock_run_command:
             load_preprocess_card_db("path_tmp", card_db, "load")
             mock_run_command.assert_called_once_with(
                 ["rgi", "load", "--card_json", str(card_db), "--local"],
@@ -22,7 +22,7 @@ class TestAnnotateReadsCARD(TestPluginBase):
 
     def test_preprocess_card_db(self):
         card_db = CARDDatabaseFormat()
-        with patch("q2_amr.utils.run_command") as mock_run_command:
+        with patch("q2_amr.card.utils.run_command") as mock_run_command:
             load_preprocess_card_db("path_tmp", card_db, "preprocess")
             mock_run_command.assert_called_once_with(
                 ["rgi", "card_annotation", "-i", str(card_db)], "path_tmp", verbose=True
@@ -30,7 +30,7 @@ class TestAnnotateReadsCARD(TestPluginBase):
 
     def test_load_card_db_fasta(self):
         card_db = self.get_data_path("card_test.json")
-        with patch("q2_amr.utils.run_command") as mock_run_command:
+        with patch("q2_amr.card.utils.run_command") as mock_run_command:
             load_preprocess_card_db("path_tmp", card_db, "load_fasta")
             mock_run_command.assert_called_once_with(
                 [
@@ -46,7 +46,7 @@ class TestAnnotateReadsCARD(TestPluginBase):
                 verbose=True,
             )
 
-    @patch("q2_amr.utils.run_command")
+    @patch("q2_amr.card.utils.run_command")
     def test_exception_raised(self, mock_run_command):
         mock_run_command.side_effect = subprocess.CalledProcessError(1, "cmd")
         tmp = "path/to/tmp"
